cleaner test suite
